🐛 fix コードレビュー §2 の不具合を修正（シグネチャ変更なし）
code_review260924.md §2 のうち、public API のシグネチャを変えずに直せるものを修正。
（§2-3 は §4-1 で対応。§2-18 / §2-19 はフェーズ5で対象コードごと削除済み）

- 2-1 LoginUser.SetUser: 別のユーザーに切り替えるときに前のユーザーの科・医師・氏名が
  残る問題を修正。pat.csv の値を引き継ぐ ReadPatCSV（同じ ID で呼ぶ）の挙動を変えないよう、
  ID が変わるときだけ Clear() する
- 2-2 Env: 別のプロパティを読むと init() が設定済みの値を初期値に戻す問題を修正。
  遅延初期化をやめ、自動プロパティの初期値にした
- 2-4 EyeOpe.GetList: 空白が2つ続くと術式・医師の条件が '%%' になり全件ヒットする問題を修正
- 2-5 PatBase: カナの全角空白の判定で name を見ていた誤りを修正
- 2-6 BaseInfo: ITEM_TITLE2 を Name1 に上書きしていた誤りを Name2 に修正
- 2-7 NidekARK1.FileSave: 固定の c:\transfile\data を見てフォルダ作成を判断していたため、
  出力先を変えるとフォルダが作られない問題を修正
- 2-8 NidekARK1.SaveDate / SaveTime: 短いファイル名で ArgumentOutOfRangeException になる問題を修正
- 2-9 AppString.ConcatLists: 空の要素で "1,,2" を作る問題と、max 引数が無視される問題を修正
- 2-10 TableData(DataGridView): 値が null のセルと新規行で NullReferenceException になる問題を修正
- 2-11 CanonRKF1Form: 受信データが空のとき例外メッセージがデータとして保存される問題を修正
- 2-12 DateTimeAgent.DateFormat: "00000000" のような不正な日付で FormatException になる問題を修正
  （LONG / SHORT は文字列操作のみで例外にならないため従来どおり）
- 2-13 EyeOpeRsv.Save: スタックトレースを失うだけの catch { throw ex; } を削除
- 2-14 EyeLensMeter.ReadData: 空ファイルで落ちる問題を列数の判定で修正
- 2-15 EyeKensaItemMaster.ListByKensaId: 存在しない検査IDで落ちる問題を修正
- 2-16 LibSettings.Read: 削除済みの SOAP 機能用に、設定を読むたびに一時フォルダの中身を
  削除していた処理を削除
- 2-17 StdControlPat1: 患者IDの判定を double.TryParse から long.TryParse に変更

確認: ML(Release/Debug)・EyeCenter・NidekARK1・CanonRKF1 のビルド成功、
EyeCenter.Tests 95件合格（画面の手動確認は未実施）

+++ Boundary/StdControlPat1.cs
@@ -173,9 +173,9 @@ namespace MedicalLibrary.Boundary
             {
                 if (this.PatIdBox1.Text.Length > 0)
                 {
-                    double pid = 0;
+                    long pid = 0;
 
-                    if (!double.TryParse(this.PatIdBox1.Text, out pid))
+                    if (!long.TryParse(this.PatIdBox1.Text, out pid))
                     {
                         MessageBox.Show("患者コードは数字のみで入力してください");
                         this.PatIdBox1.Clear();
